+++ src/jvmMain/kotlin/com/bkahlert/kommons/debug/CurrentPlatform.kt
@@ -1,0 +1,3 @@
+package com.bkahlert.kommons.debug
+
+internal actual val CurrentPlatform: Platform get() = Platform.JVM