+++ src/jvmTest/kotlin/com/bkahlert/kommons/debug/PlatformsTest.kt
@@ -1,0 +1,12 @@
+package com.bkahlert.kommons.debug
+
+import io.kotest.matchers.shouldBe
+import kotlin.test.Test
+
+class PlatformsTest {
+
+    @Test
+    fun test_current() {
+        Platform.Current shouldBe Platform.JVM
+    }
+}